Staging

See merge request aloes/aloes-client!5

+++ src/services/socket.js
@@ -1,35 +1,32 @@
+/* Copyright 2019 Edouard Maleix, read LICENSE */
+
 import mqtt from 'async-mqtt';
 import logger from './logger';
 import PubSub from './PubSub';
 
 const Storage = window.sessionStorage;
-
 const brokerUrl = process.env.VUE_APP_BROKER_URL;
+const socket = { failureCount: 0, maxFailureCount: 10 };
+
 const baseOptions = {
-  //  keepalive: 60,
-  // reschedulePings: true,
+  keepalive: 60,
+  reschedulePings: true,
   protocolId: 'MQTT',
   protocolVersion: 4,
-  reconnectPeriod: 3000,
-  connectTimeout: 30 * 1000,
+  reconnectPeriod: 2000,
+  connectTimeout: 5 * 1000,
   clean: true,
   clientId: null,
   username: null,
   password: null,
 };
 
-const socket = {};
-
 const setSocketId = socketId => {
-  if (Storage) {
-    Storage.setItem('socket-id', socketId);
-  }
+  if (Storage) Storage.setItem('socket-id', socketId);
 };
 
 const delSocketId = () => {
-  if (Storage) {
-    Storage.removeItem('socket-id');
-  }
+  if (Storage) Storage.removeItem('socket-id');
 };
 
 const getSocketId = () => {
@@ -41,13 +38,15 @@ socket.setToken = token => {
   try {
     logger.publish(3, 'socket', 'setToken:req', token);
     socket.token = token;
+    const clientId = `${token.userId.toString()}-${Math.random()
+      .toString(16)
+      .substr(2, 8)}`;
     const options = {
       ...baseOptions,
-      clientId: `${token.userId.toString()}-${Math.random()
-        .toString(16)
-        .substr(2, 8)}`,
+      clientId,
       username: token.userId.toString(),
       password: token.id.toString(),
+      // will: { topic: `${clientId}/status`, payload: 'KO?', retain: false, qos: 0 },
     };
 
     socket.initSocket(options);
@@ -62,48 +61,63 @@ socket.removeToken = () => {
     delSocketId();
     delete socket.token;
     if (socket.client) {
+      // socket.client.removeAllListeners('offline');
       PubSub.removeListeners(socket.client);
-      socket.client.end();
+      socket.client.end(true);
+      delete socket.client;
     }
     logger.publish(3, 'socket', 'removeToken:res', 'success');
-    // return;
   } catch (error) {
     logger.publish(3, 'socket', 'removeToken:err', error);
-    // throw error;
   }
+};
+
+const handleMessage = (packet, cb) => {
+  PubSub.handler(packet.topic, packet.payload)
+    .then(() => {
+      cb();
+    })
+    .catch(() => {
+      cb();
+    });
 };
 
 socket.initSocket = async options => {
   try {
     logger.publish(3, 'socket', 'initSocket:req', options);
-
     let socketId = getSocketId();
     if (socketId && socketId !== null && socket.client) {
       return socket;
     }
-    socket.client = await mqtt.connectAsync(brokerUrl, options);
-    logger.publish(3, 'socket', 'onConnect', 'success');
-
-    // socket.client = mqtt.connect(brokerUrl, options);
-    // socket.client.on('connect', async state => {
-    //   logger.publish(3, 'socket', 'onConnect', state);
-    //   setSocketId(options.clientId);
-    // });
-
     setSocketId(options.clientId);
+    socket.client = await mqtt.connectAsync(brokerUrl, options);
+    socket.failureCount = 0;
+    logger.publish(3, 'socket', 'onConnect', 'success');
+    // socket.client.setMaxListeners()
     await PubSub.setListeners(socket.client, socket.token);
 
-    socket.client.on('offline', () => {
-      logger.publish(3, 'socket', 'onDisconnect', '');
-      delSocketId();
+    // socket.client.on('reconnect', () => {
+    //   logger.publish(3, 'socket', 'onReconnecting', options.clientId);
+    //   // setSocketId(options.clientId);
+    // });
+
+    socket.client.once('offline', () => {
+      logger.publish(3, 'socket', 'onDisconnect', options.clientId);
+      // delSocketId();
     });
 
-    socket.client.on('message', PubSub.handler);
+    socket.client.handleMessage = handleMessage;
 
     return socket;
   } catch (error) {
     logger.publish(3, 'socket', 'initSocket:err', error);
-    throw error;
+    delSocketId();
+    if (socket.failureCount < socket.maxFailureCount) {
+      socket.failureCount += 1;
+      return setTimeout(() => socket.initSocket(options), baseOptions.reconnectPeriod);
+    }
+    // delete socket.token;
+    return null;
   }
 };
